+++ bot/cwl/index.js
@@ -1,17 +1,17 @@
 /**
  * * This file contains the functions to interact with the CWL files in the repository
  */
-import { consola } from 'consola'
-import { logwatch } from '../utils/logwatch.js'
+import { consola } from "consola";
+import { logwatch } from "../utils/logwatch.js";
 import {
   isRepoPrivate,
   createId,
   replaceRawGithubUrl,
-} from '../utils/tools/index.js'
-import dbInstance from '../db.js'
+} from "../utils/tools/index.js";
+import dbInstance from "../db.js";
 
-const CODEFAIR_DOMAIN = process.env.CODEFAIR_APP_DOMAIN
-const { VALIDATOR_URL } = process.env
+const CODEFAIR_DOMAIN = process.env.CODEFAIR_APP_DOMAIN;
+const { VALIDATOR_URL } = process.env;
 
 /**
  * * This function gets the CWL files in the repository
@@ -22,14 +22,14 @@ const { VALIDATOR_URL } = process.env
  */
 export function getCWLFiles(context, owner, repository) {
   return new Promise((resolve, reject) => {
-    logwatch.info('Checking for CWL files in the repository...')
+    logwatch.info("Checking for CWL files in the repository...");
 
-    const cwlFiles = []
+    const cwlFiles = [];
     const cwlObject = {
       contains_cwl_files: false,
       files: [],
       removed_files: [],
-    }
+    };
 
     const searchDirectory = async function (path) {
       try {
@@ -37,36 +37,36 @@ export function getCWLFiles(context, owner, repository) {
           owner,
           path,
           repo: repository.name,
-        })
+        });
 
         for (const file of repoContent.data) {
-          if (file.type === 'file' && file.name.endsWith('.cwl')) {
-            logwatch.info(`CWL file found: ${file.name}`)
-            cwlObject.files.push(file)
+          if (file.type === "file" && file.name.endsWith(".cwl")) {
+            logwatch.info(`CWL file found: ${file.name}`);
+            cwlObject.files.push(file);
           }
-          if (file.type === 'dir') {
-            await searchDirectory(file.path)
+          if (file.type === "dir") {
+            await searchDirectory(file.path);
           }
         }
       } catch (error) {
         if (error.status === 404) {
           // Repository is empty
-          resolve(cwlObject)
-          return
+          resolve(cwlObject);
+          return;
         }
         logwatch.error(
           {
-            message: 'Error finding CWL files throughout the repository:',
+            message: "Error finding CWL files throughout the repository:",
             error,
           },
           true
-        )
-        reject(error)
+        );
+        reject(error);
       }
-    }
+    };
 
     // Call the async function and handle its promise
-    searchDirectory('')
+    searchDirectory("")
       .then(async () => {
         try {
           // Check if the db entry exists for the repository
@@ -74,26 +74,36 @@ export function getCWLFiles(context, owner, repository) {
             where: {
               repository_id: repository.id,
             },
-          })
+          });
 
           if (existingCWL && existingCWL?.contains_cwl_files) {
-            cwlObject.contains_cwl_files = existingCWL.contains_cwl_files
+            cwlObject.contains_cwl_files = existingCWL.contains_cwl_files;
+            if (cwlObject.files.length > 0) {
+              // Remove the files that are not in the cwlObject.files array
+              const cwlFilePaths = cwlObject.files.map((file) => {
+                file.path;
+              });
+
+              cwlObject.removed_files = cwlExists.files.filter((file) => {
+                return !cwlFilePaths.includes(file.path);
+              });
+            }
           }
 
-          cwlObject.contains_cwl_files = cwlObject.files.length > 0
+          cwlObject.contains_cwl_files = cwlObject.files.length > 0;
 
-          resolve(cwlObject)
+          resolve(cwlObject);
         } catch (error) {
-          console.log('Error getting CWL files:', error)
+          console.log("Error getting CWL files:", error);
           throw new Error(
-            'Error getting the CWL files: ',
+            "Error getting the CWL files: ",
             JSON.stringify(error),
             { cause: error }
-          )
+          );
         }
       })
-      .catch(reject)
-  })
+      .catch(reject);
+  });
 }
 
 /**
@@ -108,32 +118,32 @@ export async function validateCWLFile(downloadUrl) {
         file_path: downloadUrl,
       }),
       headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
       },
-      method: 'POST',
-    })
+      method: "POST",
+    });
     if (!response.ok && response.status === 400) {
-      const error = await response.json()
+      const error = await response.json();
       // consola.warn("Validation error:", error.error);
-      return [false, error.error]
+      return [false, error.error];
     }
     if (!response.ok && response.status === 500) {
       logwatch.error(
         {
-          message: 'Error validating CWL file:',
+          message: "Error validating CWL file:",
           validation_response: response,
         },
         true
-      )
-      return [false, 'Error validating CWL file']
+      );
+      return [false, "Error validating CWL file"];
     }
     if (response.ok) {
-      const data = await response.json()
-      return [true, data.output]
+      const data = await response.json();
+      return [true, data.output];
     }
   } catch (e) {
-    logwatch.error({ message: 'Error validating CWL file:', error: e }, true)
-    return [false, 'Error validating CWL file']
+    logwatch.error({ message: "Error validating CWL file:", error: e }, true);
+    return [false, "Error validating CWL file"];
   }
 }
 
@@ -153,26 +163,26 @@ export async function applyCWLTemplate(
   owner,
   context
 ) {
-  const privateRepo = await isRepoPrivate(context, owner, repository.name)
-  const identifier = createId()
-  const overallSection = `\n\n## Language Specific Standards\n\nTo make your software FAIR is it important to follow language specific standards and best practices. Codefair will check below that your code complies with applicable standards,`
-  let url = `${CODEFAIR_DOMAIN}/dashboard/${owner}/${repository.name}/view/cwl-validation`
+  const privateRepo = await isRepoPrivate(context, owner, repository.name);
+  const identifier = createId();
+  const overallSection = `\n\n## Language Specific Standards\n\nTo make your software FAIR is it important to follow language specific standards and best practices. Codefair will check below that your code complies with applicable standards,`;
+  let url = `${CODEFAIR_DOMAIN}/dashboard/${owner}/${repository.name}/view/cwl-validation`;
 
   // Delete file entries from db if they were removed from the repository
-  if (subjects.cwl.removed_files.length > 0) {
+  if (subjects.cwl?.files && subjects.cwl.removed_files.length > 0) {
     // Remove the files from the database
     const existingCWL = await dbInstance.cwlValidation.findUnique({
       where: {
         repository_id: repository.id,
       },
-    })
+    });
 
     if (existingCWL) {
       const newFiles = existingCWL.files.filter((file) => {
-        return !subjects.cwl.removed_files.includes(file.path)
-      })
+        return !subjects.cwl.removed_files.includes(file.path);
+      });
 
-      logwatch.info(`New files after removed_files filter: ${newFiles}`)
+      logwatch.info(`New files after removed_files filter: ${newFiles}`);
 
       await dbInstance.cwlValidation.update({
         data: {
@@ -181,23 +191,23 @@ export async function applyCWLTemplate(
           updated_at: newDate,
         },
         where: { repository_id: repository.id },
-      })
+      });
     }
   }
 
   // New/Modified CWL files were found, begin validation workflow
-  const cwlFiles = []
-  let validOverall = true
-  let tableContent = ''
-  let failedCount = 0
+  const cwlFiles = [];
+  let validOverall = true;
+  let tableContent = "";
+  let failedCount = 0;
   const existingCWL = await dbInstance.cwlValidation.findUnique({
     where: {
       repository_id: repository.id,
     },
-  })
+  });
 
-  if (subjects.cwl.files.length === 0) {
-    logwatch.warn(`No CWL files found in the repository, ${repository.name}`)
+  if (subjects.cwl?.files && subjects.cwl.files.length === 0) {
+    logwatch.warn(`No CWL files found in the repository, ${repository.name}`);
 
     if (existingCWL) {
       // No CWL files found in the repository, update the db
@@ -205,63 +215,63 @@ export async function applyCWLTemplate(
         data: {
           contains_cwl_files: false,
           files: [],
-          overall_status: '',
+          overall_status: "",
         },
         where: { repository_id: repository.id },
-      })
+      });
     }
-    return baseTemplate
+    return baseTemplate;
   }
 
-  logwatch.start('Validating CWL files for', repository.name)
+  logwatch.start("Validating CWL files for", repository.name);
   // Validate each CWL file from list\
-  logwatch.info(`Validating ${JSON.stringify(subjects.cwl)} CWL files`)
+  logwatch.info(`Validating ${JSON.stringify(subjects.cwl)} CWL files`);
   if (subjects.cwl.files.length > 0) {
     for (const file of subjects.cwl.files) {
-      const fileSplit = file.name.split('.')
+      const fileSplit = file.name.split(".");
 
-      if (fileSplit.includes('cwl')) {
+      if (fileSplit.includes("cwl")) {
         const downloadUrl =
           file?.commitId && !privateRepo
-            ? file.download_url.replace('/main/', `/${file.commitId}/`)
-            : file.download_url // Replace the branch with the commit id if commit id is available and the repo is public
+            ? file.download_url.replace("/main/", `/${file.commitId}/`)
+            : file.download_url; // Replace the branch with the commit id if commit id is available and the repo is public
 
         const [isValidCWL, validationMessage] =
-          await validateCWLFile(downloadUrl)
+          await validateCWLFile(downloadUrl);
 
         if (!isValidCWL && validOverall) {
           // Overall status of CWL validations is invalid
-          validOverall = false
+          validOverall = false;
         }
 
         if (!isValidCWL) {
-          failedCount += 1
+          failedCount += 1;
         }
 
         const [modifiedValidationMessage, lineNumber1, lineNumber2] =
-          replaceRawGithubUrl(validationMessage, downloadUrl, file.html_url)
+          replaceRawGithubUrl(validationMessage, downloadUrl, file.html_url);
 
         // Add the line numbers to the URL if they exist
         if (lineNumber1) {
-          file.html_url += `#L${lineNumber1}`
+          file.html_url += `#L${lineNumber1}`;
           if (lineNumber2) {
-            file.html_url += `-L${lineNumber2}`
+            file.html_url += `-L${lineNumber2}`;
           }
         }
 
         // Create a new object for the file entry to be added to the db
-        const newDate = Math.floor(Date.now() / 1000)
+        const newDate = Math.floor(Date.now() / 1000);
         cwlFiles.push({
           href: file.html_url,
           last_modified: newDate,
           last_validated: newDate,
           path: file.path,
           validation_message: modifiedValidationMessage,
-          validation_status: isValidCWL ? 'valid' : 'invalid',
-        })
+          validation_status: isValidCWL ? "valid" : "invalid",
+        });
 
         // Apply the validation file count to the analytics collection on the db
-        const analyticsCollection = dbInstance.analytics
+        const analyticsCollection = dbInstance.analytics;
         await analyticsCollection.upsert({
           create: {
             cwl_validated_file_count: 1, // Start count at 1 when creating
@@ -275,14 +285,14 @@ export async function applyCWLTemplate(
           where: {
             id: repository.id,
           },
-        })
+        });
 
         // Add the file to the table content of the issue dashboard
-        tableContent += `| ${file.path} | ${isValidCWL ? '✔️' : '❌'} |\n`
+        tableContent += `| ${file.path} | ${isValidCWL ? "✔️" : "❌"} |\n`;
 
         logwatch.success(
-          `File: ${file.path} is ${isValidCWL ? 'valid' : 'invalid'}`
-        )
+          `File: ${file.path} is ${isValidCWL ? "valid" : "invalid"}`
+        );
       }
     }
   }
@@ -294,43 +304,43 @@ export async function applyCWLTemplate(
         contains_cwl_files: subjects.cwl.contains_cwl_files,
         files: cwlFiles,
         identifier,
-        overall_status: validOverall ? 'valid' : 'invalid',
+        overall_status: validOverall ? "valid" : "invalid",
         repository: {
           connect: {
             id: repository.id,
           },
         },
       },
-    })
+    });
 
     if (!cwlFiles.length > 0) {
       logwatch.warn(
         `No CWL files found in the repository, ${repository.name}, skipping CWL section`
-      )
-      return baseTemplate
+      );
+      return baseTemplate;
     }
   } else {
     // An entry exists in the db, thus possible old files exist (merge both lists)
-    validOverall = true
-    const fileMap = new Map()
+    validOverall = true;
+    const fileMap = new Map();
 
     // Add existing files to the map
     existingCWL.files.forEach((file) => {
-      fileMap.set(file.path, file)
-    })
+      fileMap.set(file.path, file);
+    });
     // Add new files to the map, replacing any existing entries with the same path
     cwlFiles.forEach((file) => {
-      fileMap.set(file.path, file)
-    })
+      fileMap.set(file.path, file);
+    });
 
     // Convert the map back to an array
-    const newFiles = Array.from(fileMap.values())
+    const newFiles = Array.from(fileMap.values());
 
     // Check if the overall status is still valid
     for (const file of newFiles) {
-      if (file.validation_status === 'invalid') {
-        validOverall = false
-        break
+      if (file.validation_status === "invalid") {
+        validOverall = false;
+        break;
       }
     }
 
@@ -338,43 +348,43 @@ export async function applyCWLTemplate(
       data: {
         contains_cwl_files: newFiles.length > 0,
         files: [...newFiles],
-        overall_status: validOverall ? 'valid' : 'invalid',
+        overall_status: validOverall ? "valid" : "invalid",
       },
       where: { repository_id: repository.id },
-    })
+    });
 
     if (!newFiles.length > 0) {
       // All CWL files were removed from the repository
       logwatch.warn(
         `All CWL files were removed from: ${repository.name}, skipping CWL section`
-      )
-      return baseTemplate
+      );
+      return baseTemplate;
     } else {
       // Recreate the table content to include the new and old cwl files
       logwatch.start(
-        'Recreating the table content for the CWL section to include new and old files'
-      )
-      tableContent = ''
-      failedCount = 0
+        "Recreating the table content for the CWL section to include new and old files"
+      );
+      tableContent = "";
+      failedCount = 0;
       newFiles.forEach((file) => {
-        if (file.validation_status === 'invalid') {
-          failedCount += 1
+        if (file.validation_status === "invalid") {
+          failedCount += 1;
         }
 
-        if (file.validation_status === 'invalid' && validOverall) {
-          validOverall = false
+        if (file.validation_status === "invalid" && validOverall) {
+          validOverall = false;
         }
 
-        tableContent += `| ${file.path} | ${file.validation_status === 'valid' ? '✔️' : '❌'} |\n`
-      })
+        tableContent += `| ${file.path} | ${file.validation_status === "valid" ? "✔️" : "❌"} |\n`;
+      });
     }
 
-    subjects.cwl.files = newFiles // okay to replace at this stage, used to just get the length of the new and old files for the dashboard
+    subjects.cwl.files = newFiles; // okay to replace at this stage, used to just get the length of the new and old files for the dashboard
   }
 
-  const cwlBadge = `[![CWL](https://img.shields.io/badge/View_CWL_Report-0ea5e9.svg)](${url})`
-  baseTemplate += `${overallSection}\n\n### CWL Validations ${validOverall ? '✔️' : '❗'}\n\nCodefair has detected that you are following the Common Workflow Language (CWL) standard to describe your command line tool. Codefair ran the [cwltool validator](https://cwltool.readthedocs.io/en/latest/) and ${validOverall ? `all ***${subjects.cwl.files.length}*** CWL file(s) in your repository are valid.` : `***${failedCount}/${subjects.cwl.files.length}*** CWL file(s) in your repository are not valid.`}\n\n<details>\n<summary>Summary of the validation report</summary>\n\n| File | Validation result |\n| :---- | :----: |\n${tableContent}</details>\n\nTo view the full report of each CWL file or to rerun the validation, click the "View CWL Report" button below.\n\n${cwlBadge}`
+  const cwlBadge = `[![CWL](https://img.shields.io/badge/View_CWL_Report-0ea5e9.svg)](${url})`;
+  baseTemplate += `${overallSection}\n\n### CWL Validations ${validOverall ? "✔️" : "❗"}\n\nCodefair has detected that you are following the Common Workflow Language (CWL) standard to describe your command line tool. Codefair ran the [cwltool validator](https://cwltool.readthedocs.io/en/latest/) and ${validOverall ? `all ***${subjects.cwl.files.length}*** CWL file(s) in your repository are valid.` : `***${failedCount}/${subjects.cwl.files.length}*** CWL file(s) in your repository are not valid.`}\n\n<details>\n<summary>Summary of the validation report</summary>\n\n| File | Validation result |\n| :---- | :----: |\n${tableContent}</details>\n\nTo view the full report of each CWL file or to rerun the validation, click the "View CWL Report" button below.\n\n${cwlBadge}`;
 
-  logwatch.success('CWL template section applied')
-  return baseTemplate
+  logwatch.success("CWL template section applied");
+  return baseTemplate;
 }