+++ bot/prettier.config.js
@@ -3,10 +3,10 @@
  * @type {import("prettier").Config}
  */
 const config = {
-    trailingComma: 'es5',
-    tabWidth: 2,
-    semi: false,
-    singleQuote: true,
-}
+  trailingComma: 'es5',
+  tabWidth: 2,
+  semi: true,
+  singleQuote: false,
+};
 
-export default config
+export default config;