+++ bot/utils/renderer/index.js
@@ -30,7 +30,7 @@ export async function renderIssues(
   repository,
   emptyRepo,
   subjects,
-  prInfo = { title: "", link: "" },
+  prInfo = { title: "", link: "" }
 ) {
   try {
     if (emptyRepo) {
@@ -38,67 +38,70 @@ export async function renderIssues(
         `Applying empty repo template for repository: ${repository.name}`
       );
       let emptyTemplate = `# Check the FAIRness of your software\n\nThis issue is your repository's dashboard for all things FAIR. Keep it open as making and keeping software FAIR is a continuous process that evolves along with the software. You can read the [documentation](https://docs.codefair.io/docs/dashboard.html) to learn more.\n\n> [!WARNING]\n> Currently your repository is empty and will not be checked until content is detected within your repository.\n\n## LICENSE\n\nTo make your software reusable a license file is expected at the root level of your repository. Codefair will check for a license file after you add content to your repository.\n\n![License](https://img.shields.io/badge/License_Not_Checked-fbbf24)\n\n## Metadata\n\nTo make your software FAIR a CITATION.cff and codemeta.json metadata files are expected at the root level of your repository. Codefair will check for these files after a license file is detected.\n\n![Metadata](https://img.shields.io/badge/Metadata_Not_Checked-fbbf24)`;
-  
+
       emptyTemplate = applyLastModifiedTemplate(emptyTemplate);
-  
+
       return emptyTemplate;
     }
-  
+
     let baseTemplate = `# Check the FAIRness of your software\n\nThis issue is your repository's dashboard for all things FAIR. Keep it open as making and keeping software FAIR is a continuous process that evolves along with the software. You can read the [documentation](https://docs.codefair.io/docs/dashboard.html) to learn more.\n\n`;
-  
+
     baseTemplate = await applyLicenseTemplate(
       subjects,
       baseTemplate,
       repository,
       owner,
-      context,
+      context
     );
-  
+
     // Check if pull request url exist in db
     const prUrl = await dbInstance.licenseRequest.findUnique({
       where: { repository_id: repository.id },
     });
-  
+
     if (prUrl?.license_id === "Custom") {
       subjects.customLicense = true;
     }
-  
+
     // If License PR is open, add the PR number to the dashboard
     if (prUrl?.pull_request_url !== "") {
       // Verify if the PR is still open
       try {
         const pr = await context.octokit.pulls.get({
-        owner,
-        repo: repository.name,
-        pull_number: prUrl.pull_request_url.split("/").pop(),
+          owner,
+          repo: repository.name,
+          pull_number: prUrl.pull_request_url.split("/").pop(),
         });
-  
+
         if (pr.data.state === "open") {
-        baseTemplate += `\n\nA pull request for the LICENSE file is open. You can view the pull request:\n\n[![License](https://img.shields.io/badge/View_PR-6366f1.svg)](${prUrl.pull_request_url})`;
+          baseTemplate += `\n\nA pull request for the LICENSE file is open. You can view the pull request:\n\n[![License](https://img.shields.io/badge/View_PR-6366f1.svg)](${prUrl.pull_request_url})`;
         } else {
-        // If the PR is closed, remove the PR from the database
-        await dbInstance.licenseRequest.update({
-          where: { repository_id: repository.id },
-          data: { pull_request_url: "" },
-        });
+          // If the PR is closed, remove the PR from the database
+          await dbInstance.licenseRequest.update({
+            where: { repository_id: repository.id },
+            data: { pull_request_url: "" },
+          });
         }
       } catch (error) {
-        logwatch.error({message: "Error fetching pull request:", error}, true);
+        logwatch.error(
+          { message: "Error fetching pull request:", error },
+          true
+        );
       }
     }
-  
+
     baseTemplate = await applyMetadataTemplate(
       subjects,
       baseTemplate,
       repository,
       owner,
-      context,
+      context
     );
-  
+
     const metadataPrUrl = await dbInstance.codeMetadata.findUnique({
       where: { repository_id: repository.id },
     });
-  
+
     if (metadataPrUrl?.pull_request_url) {
       try {
         const pr = await context.octokit.pulls.get({
@@ -106,7 +109,7 @@ export async function renderIssues(
           repo: repository.name,
           pull_number: metadataPrUrl.pull_request_url.split("/").pop(),
         });
-  
+
         if (pr.data.state === "open") {
           baseTemplate += `\n\nA pull request for the metadata files is open. You can view the pull request:\n\n[![Metadata](https://img.shields.io/badge/View_PR-6366f1.svg)](${metadataPrUrl.pull_request_url})`;
         } else {
@@ -117,30 +120,28 @@ export async function renderIssues(
           });
         }
       } catch (error) {
-        logwatch.error({message: "Error fetching metadata pull request:", error}, true);
+        logwatch.error(
+          { message: "Error fetching metadata pull request:", error },
+          true
+        );
       }
     }
-  
+
     baseTemplate = await applyCWLTemplate(
       subjects,
       baseTemplate,
       repository,
       owner,
-      context,
+      context
     );
-  
-    baseTemplate = await applyArchivalTemplate(
-      baseTemplate,
-      repository,
-      owner,
-    );
-  
-    baseTemplate = applyLastModifiedTemplate(baseTemplate);
-  
-    return baseTemplate;
 
+    baseTemplate = await applyArchivalTemplate(baseTemplate, repository, owner);
+
+    baseTemplate = applyLastModifiedTemplate(baseTemplate);
+
+    return baseTemplate;
   } catch (error) {
-    throw new Error("Error rendering issue:", { cause: error});
+    throw new Error("Error rendering issue:", { cause: error });
   }
 }