+++ bot/utils/tools/index.js
@@ -13,6 +13,13 @@ import dbInstance from "../../db.js";
 dayjs.extend(utc);
 dayjs.extend(timezone);
 
+const IGNORED_COMMIT_MESSAGES = [
+  "chore: 📝 Update CITATION.cff with Zenodo identifier",
+  "chore: 📝 Update codemeta.json with Zenodo identifier",
+];
+
+const { GH_APP_NAME } = process.env;
+
 /**
  * * Initialize the database connection
  * @returns {Promise<boolean>} - Returns true if the database is connected, false otherwise
@@ -24,7 +31,7 @@ export async function intializeDatabase() {
     logwatch.success("Connected to database!");
     return true;
   } catch (error) {
-    logwatch.error({message: "Error connecting to database:", error}, true);
+    logwatch.error({ message: "Error connecting to database:", error }, true);
   }
 }
 
@@ -64,7 +71,10 @@ export async function getDefaultBranch(context, owner, repositoryName) {
 
     return defaultBranch.data.default_branch;
   } catch (error) {
-    logwatch.error({message: "Error getting the default branch:", error}, true);
+    logwatch.error(
+      { message: "Error getting the default branch:", error },
+      true
+    );
   }
 }
 
@@ -159,7 +169,7 @@ export async function gatherRepoAuthors(context, owner, repo, fileType) {
       return await context.octokit.users.getByUsername({
         username: contributor.login,
       });
-    }),
+    })
   );
 
   const parsedAuthors = [];
@@ -244,7 +254,7 @@ export async function getDOI(context, owner, repoName) {
     });
 
     const readmeContent = Buffer.from(readme.data.content, "base64").toString(
-      "utf-8",
+      "utf-8"
     );
     const doiRegex = /10.\d{4,9}\/[-._;()/:A-Z0-9]+/i;
     const doi = doiRegex.exec(readmeContent);
@@ -268,11 +278,11 @@ export async function verifyRepoName(
   dbRepoName,
   repository,
   owner,
-  collection,
+  collection
 ) {
   if (dbRepoName !== repository.name) {
     logwatch.info(
-      `Repository name for ${owner} has changed from ${dbRepoName} to ${repository.name}`,
+      `Repository name for ${owner} has changed from ${dbRepoName} to ${repository.name}`
     );
 
     // Check if the installation is already in the database
@@ -306,7 +316,10 @@ export async function isRepoEmpty(context, owner, repoName) {
     if (error.status === 404) {
       return true;
     }
-    logwatch.error({message: "Error checking if the repository is empty:", error}, true);
+    logwatch.error(
+      { message: "Error checking if the repository is empty:", error },
+      true
+    );
   }
 }
 
@@ -321,13 +334,20 @@ export async function verifyInstallationAnalytics(
   context,
   repository,
   actionCount = 5,
-  latestCommitInfo = {},
+  latestCommitInfo = {}
 ) {
+  const installationId = context.payload.installation.id;
   const owner =
     context.payload?.installation?.account?.login ||
     context.payload?.repository?.owner?.login;
 
-  const installationId = context.payload.installation.id;
+  if (!installationId) {
+    throw new Error("Installation ID is missing");
+  }
+
+  if (!owner) {
+    throw new Error("Owner information is missing");
+  }
 
   const installation = await dbInstance.installation.findUnique({
     where: {
@@ -357,6 +377,7 @@ export async function verifyInstallationAnalytics(
       },
     });
   } else {
+    // If the installation is in the database, check the action count to determine if the limit has been reached
     if (installation.action_count > 0) {
       await dbInstance.installation.update({
         data: {
@@ -376,10 +397,8 @@ export async function verifyInstallationAnalytics(
     }
 
     if (installation.action_count === 0) {
-      logwatch.info(`Action limit reached for ${installation.repo}, no longer limiting actions`);
       await dbInstance.installation.update({
         data: {
-          action_count: 0,
           latest_commit_date: latestCommitInfo.latest_commit_date || "",
           latest_commit_message: latestCommitInfo.latest_commit_message || "",
           latest_commit_sha: latestCommitInfo.latest_commit_sha || "",
@@ -388,11 +407,13 @@ export async function verifyInstallationAnalytics(
         where: { id: repository.id },
       });
     }
-    verifyRepoName(
+
+    // Verify if the repository name has changed
+    await verifyRepoName(
       installation.repo,
       repository,
       owner,
-      dbInstance.installation,
+      dbInstance.installation
     );
   }
 
@@ -421,11 +442,14 @@ export async function isRepoPrivate(context, owner, repoName) {
     });
 
     logwatch.info(
-      `Repository ${repoName} is private: ${repoDetails.data.private}`,
+      `Repository ${repoName} is private: ${repoDetails.data.private}`
     );
     return repoDetails.data.private;
   } catch (error) {
-    logwatch.error({message: "Error verifying if the repository is private:", error}, true);
+    logwatch.error(
+      { message: "Error verifying if the repository is private:", error },
+      true
+    );
   }
 }
 
@@ -454,7 +478,7 @@ export function replaceRawGithubUrl(inputString, oldUrl, newUrl) {
   // Regex to find the oldUrl with optional line numbers in the format :line:column
   const urlRegex = new RegExp(
     `(${oldUrl.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")})(:\\d+:\\d+)?`,
-    "g",
+    "g"
   );
 
   let firstLineNumber = null;
@@ -495,7 +519,12 @@ export function applyLastModifiedTemplate(baseTemplate) {
  * @param {String} releaseId - ID of the release being requested
  * @returns {Object} - Release information
  */
-export async function getReleaseById(context, repositoryName, owner, releaseId) {
+export async function getReleaseById(
+  context,
+  repositoryName,
+  owner,
+  releaseId
+) {
   try {
     const draftRelease = await context.octokit.repos.getRelease({
       owner,
@@ -507,7 +536,9 @@ export async function getReleaseById(context, repositoryName, owner, releaseId) 
 
     return draftRelease;
   } catch (error) {
-    throw new Error(`Error fetching the draft release: ${error}`, { cause: error });
+    throw new Error(`Error fetching the draft release: ${error}`, {
+      cause: error,
+    });
   }
 }
 
@@ -518,7 +549,12 @@ export async function getReleaseById(context, repositoryName, owner, releaseId) 
  * @param {String} branch - Branch to download from (optional)
  * @returns {Object} - Data of the Zipball archive
  */
-export async function downloadRepositoryZip(context, owner, repositoryName, branch="") {
+export async function downloadRepositoryZip(
+  context,
+  owner,
+  repositoryName,
+  branch = ""
+) {
   try {
     if (!branch) {
       branch = await getDefaultBranch(context, owner, repositoryName);
@@ -526,12 +562,253 @@ export async function downloadRepositoryZip(context, owner, repositoryName, bran
     const { data } = await context.octokit.repos.downloadZipballArchive({
       owner,
       repo: repositoryName,
-      ref: branch
+      ref: branch,
     });
 
-    logwatch.success(`Downloaded the repository archive successfully for: ${repositoryName}`);
+    logwatch.success(
+      `Downloaded the repository archive successfully for: ${repositoryName}`
+    );
     return data;
   } catch (error) {
-    throw new Error(`Error download the repository archive for ${repositoryName}: ${error}`, { cause: error });
+    throw new Error(
+      `Error download the repository archive for ${repositoryName}: ${error}`,
+      { cause: error }
+    );
+  }
+}
+
+export async function iterateCommitDetails(commits, subjects, repository) {
+  const gatheredCWLFiles = [];
+  const removedCWLFiles = [];
+  for (let i = 0; i < commits.length; i++) {
+    if (commits[i]?.added?.length > 0) {
+      // Iterate through the added files
+      for (let j = 0; j < commits[i]?.added.length; j++) {
+        if (commits[i].added[j] === "LICENSE") {
+          subjects.license = true;
+          continue;
+        }
+        if (commits[i].added[j] === "CITATION.cff") {
+          subjects.citation = true;
+          continue;
+        }
+        if (commits[i].added[j] === "codemeta.json") {
+          subjects.codemeta = true;
+          continue;
+        }
+        const fileSplit = commits[i].added[j].split(".");
+        if (fileSplit.includes("cwl")) {
+          gatheredCWLFiles.push({
+            commitId: commits[i].id,
+            filePath: commits[i].added[j],
+          });
+          continue;
+        }
+      }
+    }
+    // Iterate through the modified files
+    if (commits[i]?.modified?.length > 0) {
+      for (let j = 0; j < commits[i]?.modified.length; j++) {
+        const fileSplit = commits[i]?.modified[j].split(".");
+        if (fileSplit.includes("cwl")) {
+          gatheredCWLFiles.push({
+            commitId: commits[i].id,
+            filePath: commits[i].modified[j],
+          });
+          continue;
+        }
+      }
+    }
+
+    // Iterate through the remove files
+    if (commits[i]?.removed?.length > 0) {
+      for (let j = 0; j < commits[i]?.removed.length; j++) {
+        const fileSplit = commits[i]?.removed[j].split(".");
+        if (fileSplit.includes("cwl")) {
+          removedCWLFiles.push(commits[i].removed[j]);
+          continue;
+        }
+        if (commits[i]?.removed[j] === "LICENSE") {
+          subjects.license = false;
+          continue;
+        }
+        if (commits[i]?.removed[j] === "CITATION.cff") {
+          subjects.citation = false;
+          continue;
+        }
+        if (commits[i]?.removed[j] === "codemeta.json") {
+          subjects.codemeta = false;
+          continue;
+        }
+      }
+    }
+  }
+
+  if (gatheredCWLFiles.length > 0) {
+    // Begin requesting the file metadata for each file name
+    for (const file of gatheredCWLFiles) {
+      const cwlFile = await context.octokit.repos.getContent({
+        owner,
+        path: file.filePath,
+        repo: repository.name,
+      });
+
+      cwlFile.data.commitId = file.commitId;
+      subjects.cwl.files.push(cwlFile.data);
+    }
+  }
+
+  subjects.cwl.contains_cwl_files = subjects.cwl.files.length > 0 || false;
+  subjects.cwl.files = subjects.cwl.files.filter(
+    (file) => !removedCWLFiles.includes(file.path)
+  );
+  subjects.cwl.removed_files = removedCWLFiles;
+
+  const cwlExists = await dbInstance.cwlValidation.findUnique({
+    where: {
+      repository_id: repository.id,
+    },
+  });
+
+  // Does the repository already contain CWL files
+  if (cwlExists) {
+    subjects.cwl.contains_cwl_files = cwlExists.contains_cwl_files;
+  }
+
+  return subjects;
+}
+
+export async function ignoreCommitMessage(commitMessage, author) {
+  if (
+    IGNORED_COMMIT_MESSAGES.includes(commitMessage) &&
+    author === `${GH_APP_NAME}[bot]`
+  ) {
+    logwatch.info(
+      `Ignoring commit message: ${commitMessage} by ${author} as it is a known commit message`
+    );
+    return true;
+  }
+  return false;
+}
+
+export async function gatherCommitDetails(context, owner, repository) {
+  // Get the name of the main branch
+  const mainBranch = await getDefaultBranch(context, owner, repository.name);
+  // Gather the latest commit to main info
+  const latestCommit = await context.octokit.repos.getCommit({
+    owner,
+    ref: mainBranch,
+    repo: repository.name,
+  });
+
+  return {
+    latest_commit_sha: latestCommit.data.sha || "",
+    latest_commit_message: latestCommit.data.commit.message || "",
+    latest_commit_url: latestCommit.data.html_url || "",
+    latest_commit_date: latestCommit.data.commit.committer.date || "",
+  };
+}
+
+export async function purgeDBEntry(repository) {
+  // Check if the installation is already in the database
+  const installation = await dbInstance.installation.findUnique({
+    where: {
+      id: repository.id,
+    },
+  });
+
+  const license = await dbInstance.licenseRequest.findUnique({
+    where: {
+      repository_id: repository.id,
+    },
+  });
+
+  const metadata = await dbInstance.codeMetadata.findUnique({
+    where: {
+      repository_id: repository.id,
+    },
+  });
+
+  const cwl = await dbInstance.cwlValidation.findUnique({
+    where: {
+      repository_id: repository.id,
+    },
+  });
+
+  const zenodoDeposition = await dbInstance.zenodoDeposition.findUnique({
+    where: {
+      repository_id: repository.id,
+    },
+  });
+
+  if (license) {
+    await dbInstance.licenseRequest.delete({
+      where: {
+        repository_id: repository.id,
+      },
+    });
+  }
+
+  if (metadata) {
+    await dbInstance.codeMetadata.delete({
+      where: {
+        repository_id: repository.id,
+      },
+    });
+  }
+
+  if (cwl) {
+    await dbInstance.cwlValidation.delete({
+      where: {
+        repository_id: repository.id,
+      },
+    });
+  }
+
+  if (zenodoDeposition) {
+    await dbInstance.zenodoDeposition.delete({
+      where: {
+        repository_id: repository.id,
+      },
+    });
+  }
+
+  if (installation) {
+    // Remove from the database
+    await dbInstance.installation.delete({
+      where: {
+        id: repository.id,
+      },
+    });
+  }
+
+  logwatch.info(`Repository uninstalled: ${repository.name}`);
+}
+
+export async function disableCodefairOnRepo(context) {
+  const { repository } = context.payload;
+  const installation = await dbInstance.installation.findUnique({
+    where: {
+      id: repository.id,
+    },
+  });
+
+  // Update installation table to disable the repository
+  if (installation) {
+    await dbInstance.installation.update({
+      data: { disabled: true },
+      where: { id: repository.id },
+    });
+  }
+
+  // If the action was just closing the issue, update the issue body
+  if (context.payload.action === "closed") {
+    // Update the body of the issue to reflect that the repository is disabled
+    await context.octokit.issues.update({
+      body: CLOSED_ISSUE_BODY,
+      issue_number: context.payload.issue.number,
+      owner: repository.owner.login,
+      repo: repository.name,
+    });
   }
 }