+++ ui/pages/dashboard/[owner]/[repo]/index.vue
@@ -21,6 +21,9 @@ const botNotInstalled = ref(false);
 const cwlValidationRerunRequestLoading = ref(false);
 const displayMetadataValidationResults = ref(false);
 const showModal = ref(false);
+const showLicenseModal = ref(false);
+const showMetadataModal = ref(false);
+const loading = ref(false);
 
 const renderIcon = (icon: string) => {
   return () => {
@@ -96,6 +99,13 @@ const hideConfirmation = () => {
 
 const showConfirmation = () => {
   showModal.value = true;
+};
+
+const handlePositiveClick = async (reRunType: string) => {
+  loading.value = true;
+  await rerunCodefairChecks(reRunType);
+  loading.value = false;
+  showLicenseModal.value = false;
 };
 
 const rerunCwlValidation = async () => {
@@ -202,10 +212,10 @@ const handleSettingsSelect = (key: any) => {
     }
   } else if (key === "re-validate-license") {
     // rerunCodefairChecks("license");
-    showConfirmation();
+    showLicenseModal.value = true;
   } else if (key === "re-validate-metadata") {
     // rerunCodefairChecks("metadata");
-    showConfirmation();
+    showMetadataModal.value = true;
   }
 };
 </script>
@@ -309,21 +319,24 @@ const handleSettingsSelect = (key: any) => {
             </n-dropdown>
 
             <n-modal
-              v-model:show="showModal"
+              v-model:show="showLicenseModal"
               :mask-closable="false"
               preset="dialog"
               title="Are you sure?"
               content="Doing this action will overwrite any existing draft. Do you want to continue?"
               positive-text="Confirm"
               negative-text="Cancel"
-              @positive-click="rerunCodefairChecks('license')"
-              @negative-click="hideConfirmation"
+              @positive-click="handlePositiveClick('license')"
+              @negative-click="showLicenseModal = false"
+              :loading="loading"
             />
           </div>
         </template>
 
         <template #content>
-          <p class="text-base">A License is required according to the FAIR-BioRS guidelines</p>
+          <p class="text-base">
+            A License is required according to the FAIR-BioRS guidelines
+          </p>
         </template>
 
         <template #action>
@@ -436,15 +449,16 @@ const handleSettingsSelect = (key: any) => {
             </n-dropdown>
 
             <n-modal
-              v-model:show="showModal"
+              v-model:show="showMetadataModal"
               :mask-closable="false"
               preset="dialog"
               title="Are you sure?"
               content="Doing this action will overwrite any existing draft. Do you want to continue?"
               positive-text="Confirm"
               negative-text="Cancel"
-              @positive-click="rerunCodefairChecks('metadata')"
-              @negative-click="hideConfirmation"
+              @positive-click="handlePositiveClick('metadata')"
+              @negative-click="showMetadataModal = false"
+              :loading="loading"
             />
           </n-flex>
         </template>